backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "porkchop-backend"
version = "0.1.0"
description = "Backend for porkchop workflow validation system using Ollama LLM"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.116.1",
    "uvicorn==0.35.0",
    "sqlalchemy==2.0.43",
    "alembic==1.16.4",
    "pydantic==2.11.7",
    "python-multipart==0.0.20",
    "ollama==0.5.3",
    "python-dotenv==1.1.1",
    "aiofiles==24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.1",
    "pytest-asyncio==1.1.0",
    "pytest-mock==3.14.0",
    "pytest-cov==6.2.1",
    "httpx==0.28.1",
    "faker==37.5.3",
    "freezegun==1.5.1",
    "ruff==0.12.9",
    "pre-commit==4.3.0"
]

[tool.ruff]
target-version = "py311"
line-length = 88
lint.select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
lint.ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument tabs
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = [
    "-s",
    "-l",
    "-vv",
    "--tb=short",
]
markers = [
    "unit: 単体テスト",
    "integration: 統合テスト",
    "e2e: エンドツーエンドテスト",
    "slow: 実行に時間がかかるテスト",
]
